--- a/src/libraries/System.Private.CoreLib/src/System/Threading/Tasks/Task.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Threading/Tasks/Task.cs
@@ -2851,6 +2851,11 @@ namespace System.Threading.Tasks
             bool returnValue = SpinWait(millisecondsTimeout);
             if (!returnValue)
             {
+                if (Thread.CurrentThread.IsThreadPoolThread)
+                {
+                    ThreadPool.RecordBlockingCallsite();
+                }
+
                 var mres = new SetOnInvokeMres();
                 try
                 {