--- a/src/mono/netcore/System.Private.CoreLib/src/System/Delegate.Mono.cs
+++ b/src/mono/netcore/System.Private.CoreLib/src/System/Delegate.Mono.cs
@@ -533,7 +533,7 @@ namespace System
             return (m != null ? m.GetHashCode() : GetType().GetHashCode()) ^ RuntimeHelpers.GetHashCode(_target);
         }
 
-        protected virtual MethodInfo GetMethodImpl()
+        protected internal virtual MethodInfo GetMethodImpl()
         {
             if (method_info != null)
                 return method_info;