Implementation of a technique; developed with @davidfowl, to prevent total `ThreadPool` starvation due to blocking Tasks on ThreadPool threads.

Main aim is to reduce chance of the process becoming irrecoverably unresponsive from "sync-over-async" in the ThreadPool; rather than making blocking on the ThreadPool a performant option.

Mechanism:

1. First time a ThreadPool thread blocks; switch on work item callsite monitoring
2. Subsequent times record the callsite as potentially blocking
3. When queuing/executing a potentially blocking callsite use a limited concurrency queue*; which is processed as a `IThreadPoolWorkItem`
4. When the blocking has subsided the marked callsites are unmarked, and blocking mitigation is disabled.

*Concurrency is limited to `ThreadCount - 1` so there is always the availability of a ThreadPool thread for non-blocking work.

This allows non-blocking items to continue executing on the ThreadPool rather than be also blocked waiting for the blocking items to complete.

Test [gist](https://gist.github.com/benaadams/9290257de1c907b3ecafc750779e2621); also added as ThreadPool test.
```
Without mitigation -> Time: 1 hr 59 min 53.12 sec
With mitigation    -> Time:              0.56 sec
```

/cc @kouvel @stephentoub