@benaadams 2021-01-23
Test ASP.NET Core endpoints:
```csharp
public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
{
    app.UseRouting();

    app.UseEndpoints(endpoints =>
    {
        endpoints.MapGet("/blocktp", context =>
        {
            var random = new Random();
            Task.Delay(random.Next(50, 500)).Wait();
            context.Response.WriteAsync("Hello World!").Wait();
            return Task.CompletedTask;
        });
        endpoints.MapGet("/async", async context =>
        {
            var random = new Random();
            await Task.Delay(random.Next(50, 500));
            await context.Response.WriteAsync("Hello World!");
        });
    });
}
```

Without the mitigation
```
$ wrk -c 128 http://.../blocktp
Running 10s test @ http://.../blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B

$ wrk -c 128 http://.../async
Running 10s test @ http://.../async
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B
```
With the mitigation
```
$ wrk -c 128 http://.../blocktp
Running 10s test @ http://...blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.00s   565.36ms   1.94s    56.41%
    Req/Sec    13.54      8.87    50.00     83.58%
  219 requests in 10.02s, 26.31KB read
  Socket errors: connect 0, read 0, write 0, timeout 180
Requests/sec:     21.86
Transfer/sec:      2.63KB

$ wrk -c 128 http://.../async
Running 10s test @ http://.../async
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   284.69ms  129.63ms 518.02ms   57.58%
    Req/Sec   220.79     51.76   343.00     75.00%
  4404 requests in 10.02s, 529.00KB read
Requests/sec:    439.70
Transfer/sec:     52.82KB
```