@stephentoub 2021-01-23
I'll take a look at the code on Monday,  but just from your description,  lots of warning bells going off in my head ;-)