@benaadams 2021-01-23
>  lots of warning bells going off in my head 

Not denying that 😅 

Mostly reduces the chance of becoming irrecoverable; rather than making blocking on the ThreadPool a performant option.

e.g. currently if 3 blocking tests are run in sequence they all fail; as does an async one after: 
```
~$ wrk -c 128 http://...:5000/blocktp
Running 10s test @ http://...:5000/blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B

~$ wrk -c 128 http://...:5000/blocktp
Running 10s test @ http://...:5000/blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B

~$ wrk -c 128 http://...:5000/blocktp
Running 10s test @ http://...:5000/blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B

~$ wrk -c 128 http://...:5000/async
Running 10s test @ http://...:5000/async
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B
```
Whereas after; still get timeouts for the blocking and fails at the start (as all threads start blocked); but it then recovers and the async still runs
```
~$ wrk -c 128 http://...:5000/blocktp
Running 10s test @ http://...:5000/blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.02s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B

~$ wrk -c 128 http://...:5000/blocktp
Running 10s test @ http://...:5000/blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec    29.11     20.92    80.00     75.41%
  192 requests in 10.02s, 23.06KB read
  Socket errors: connect 0, read 0, write 0, timeout 192
Requests/sec:     19.16
Transfer/sec:      2.30KB

~$ wrk -c 128 http://...:5000/blocktp
Running 10s test @ http://...:5000/blocktp
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.82s   145.96ms   2.00s    85.07%
    Req/Sec    37.81     27.12   140.00     59.31%
  567 requests in 10.02s, 68.11KB read
  Socket errors: connect 0, read 0, write 0, timeout 232
Requests/sec:     56.61
Transfer/sec:      6.80KB

~$ wrk -c 128 http://...:5000/async
Running 10s test @ http://...:5000/async
  2 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   279.53ms  129.31ms 513.32ms   58.99%
    Req/Sec   224.29     48.89   340.00     72.50%
  4475 requests in 10.02s, 537.52KB read
Requests/sec:    446.51
Transfer/sec:     53.63KB
```