@KalleOlaviNiemitalo 2021-01-23
Would the tracking of call sites interfere with AssemblyLoadContext.Unload?