@davidfowl 2021-01-23
This looks really promising. I think the reflection part needs to be solved but the idea that we would attempt to educate the thread pool about blocking via calls that pass through the the BCL itself!