@benaadams 2021-01-23
> I think the reflection part needs to be solved 

Solved... Forgot it was the runtime and could just add `internal` to somethings 😅