@davidfowl 2021-01-24
One problematic area are work items that batch items (basically hiding this knowledge from the thread pool. Like the IOQueue in kestrel and the Linux sockets implementation). The one to many relationship means that one bad inner work item causes all of them to be put into the blocking list.

We'd need to detect these types of work items or give them a way to opt out or participate themselves in blocking detection