@benaadams 2021-01-24
```csharp
static async Task Main(string[] args)
{
    const int count = 1_000;

    var sw = Stopwatch.StartNew();

    List<Task> tasks = new List<Task>(count);
    for (int i = 0; i < count; i++)
    {
        tasks.Add(Task.Run(() => Blocking()));
    }

    await Task.WhenAll(tasks);

    sw.Stop();

    Console.WriteLine($"Time: {sw.Elapsed}");
}

static void Blocking()
{
    Task.Delay(1).Wait();
}
```

Current: `2 mins 23.3 sec`
PR: `9.39 sec`