@benaadams 2021-01-24
```csharp
const int count = 2_500;

Action action = static () => Blocking();

// Trigger recording
await Task.Run(action);
// Trigger mitigation
await Task.Run(action);

List<Task> tasks = new List<Task>(count);
for (int i = 0; i < count; i++)
{
    tasks.Add(Task.Run(action));
}

var sw = Stopwatch.StartNew();

await Task.WhenAll(tasks);

sw.Stop();

Console.WriteLine($"Time: {sw.Elapsed}");

static void Blocking()
{
    Task.Delay(1).Wait();
}
```
`Time: 00:00:57.9761664` <- Current
`Time: 00:00:07.8701218` <- PR