@benaadams 2021-01-24
Added test (from https://gist.github.com/benaadams/9290257de1c907b3ecafc750779e2621)
```
Without mitigation -> Time: 01:59:53.1175255
With mitigation    -> Time: 00:00:00.5634481
```